merged data and set charts.
Need to work on making the words bigger on the charts. Also one of my charts in plotting in a mixed order so I still need to fix it. I have one more chart to plot and then I should be done. Let me know if you have any questions. Thanks